+++ packages/eslint-config-airbnb-base/rules/variables.js
@@ -1,3 +1,5 @@
+const restrictedGlobals = require('eslint-restricted-globals');
+
 module.exports = {
   rules: {
     // enforce or disallow variable initializations at definition
@@ -14,7 +16,7 @@ module.exports = {
     'no-label-var': 'error',
 
     // disallow specific globals
-    'no-restricted-globals': 'off',
+    'no-restricted-globals': ['error'].concat(restrictedGlobals),
 
     // disallow declaration of variables already declared in the outer scope
     'no-shadow': 'error',
@@ -34,9 +36,9 @@ module.exports = {
     'no-undefined': 'off',
 
     // disallow declaration of variables that are not used in the code
-    'no-unused-vars': ['error', { vars: 'local', args: 'after-used' }],
+    'no-unused-vars': ['error', { vars: 'all', args: 'after-used', ignoreRestSiblings: true }],
 
     // disallow use of variables before they are defined
-    'no-use-before-define': 'error'
+    'no-use-before-define': ['error', { functions: true, classes: true, variables: true }],
   }
 };